add client folder structure, abstract Home page
